Added error message (checkFinished()) functions
Polished some things and added functions for TeamMapData that is used in
Map class to check all items in the teamsMapData HashMap to see if every
data element is defined in TeamMapData and otherwise sets isFinished to
false and gives an error string.

To-Do: create a format for mapYML.yml(s) and code the writeConfig method that writes all of the data and parameters from the Map object class into a .yml file named after the world name.