Restructuring Map Classes teamsMapData Attribute
Changed the teamsMapData from a HashMap<TeamData, TeamMapData> to an
ArrayList<TeamMapData> as TeamMapData objects have a TeamData attribute.
The previous structure didn't make sense since I was just looping
through the HashMap every time it was used anyways. Also added an
ArrayList<TeamData> allTeams attribute for use of teamsMapData object
when looping through.

Added ShopType enum and TeamUpgrades (for team shop) to the structure of
the whole plugin.

TO-DO:
  * Writing teams to .yml files
  * Test full writing to config
  * Reading from .yml files to Map object